Reformatted pages and components for login and sign up
