+++ theBookCooperUI/src/components/Signup.js
@@ -1,40 +1,56 @@
 // src/components/Signup.js
-import React, { useState } from 'react';
-import { auth } from '../firebase-config';
+import {useState} from "react";
+import {Link, useNavigate} from "react-router-dom";
+import {getAuth, createUserWithEmailAndPassword} from "firebase/auth";
 
-function Signup() {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
 
-  const signup = async (e) => {
-    e.preventDefault();
-    try {
-      const userCredential = await auth.createUserWithEmailAndPassword(email, password);
-      console.log("Account created:", userCredential.user);
-    } catch (error) {
-      console.error("Error signing up:", error.message);
-    }
-  };
+const Signup = () => {
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const [confirmPassword, setConfirmPassword] = useState("");
+    const [error, setError] = useState("");
 
-  return (
-    <div>
-      <form onSubmit={signup}>
-        <input
-          type="email"
-          value={email}
-          onChange={(e) => setEmail(e.target.value)}
-          placeholder="Email"
-        />
-        <input
-          type="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          placeholder="Password"
-        />
-        <button type="submit">Sign Up</button>
-      </form>
-    </div>
-  );
+    const navigate = useNavigate();
+
+    const createAccount = async () => {
+        try {
+            if (password !== confirmPassword) {
+                setError("Password and confirm password do not match");
+                return;
+            }
+
+            await createUserWithEmailAndPassword(getAuth(), email, password);
+            navigate("/");
+        } catch(e) {
+            setError(e.message);
+        }
+    };
+
+    return (
+        <>
+            <h1>Create Account</h1>
+            {error && <p className="error">{error}</p>}
+            <input
+                placeholder="Your email address"
+                value={email}
+                onChange={e=>setEmail(e.target.value)}
+            />
+            <input
+                type="password"
+                placeholder="Your password"
+                value={password}
+                onChange={e=>setPassword(e.target.value)}
+            />
+            <input
+                type="password"
+                placeholder="Re-enter your password"
+                value={confirmPassword}
+                onChange={e=>setConfirmPassword(e.target.value)}
+            />
+            <button onClick={createAccount}>Log In</button> <br/> <br/>
+            <Link to="/login">Already have an account? Log in here.</Link>
+        </>
+    );
 }
 
 export default Signup;
