+++ theBookCooperUI/src/firebase-config.js
@@ -1,6 +1,6 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
-import { getAnalytics } from "firebase/analytics";
+//import { getAnalytics } from "firebase/analytics";
 import { getAuth } from "firebase/auth"; // Import Authentication
 
 // Your web app's Firebase configuration
@@ -18,7 +18,7 @@ const firebaseConfig = {
 const app = initializeApp(firebaseConfig);
 
 // Initialize Firebase Analytics
-const analytics = getAnalytics(app);
+//const analytics = getAnalytics(app);
 
 // Initialize Firebase Authentication and export it
 export const auth = getAuth(app);
