+++ theBookCooperUI/src/pages/SignupPage.js
@@ -1,0 +1,13 @@
+import Header from '../components/Header';
+import Signup from "../components/Signup";
+
+const SignupPage = () => {
+    return (
+        <div>
+            <Header />
+            <Signup />
+        </div>
+    );
+}
+
+export default SignupPage;
